# neat_insight/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neat Insight API",
    "version": "1.0.0",
...
    "license": {
      "name": "Apache License 2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "The current Insight installation"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/peripherals": {
      "get": {
        "tags": ["Peripherals"],
        "summary": "Get the current peripheral catalog",
        "description": "Reads the selected board's authoritative daemon catalog. Insight preserves the generic v1 envelope and device records and adds selected-board context; it never performs a backup scan.",
        "operationId": "getPeripheralCatalog",
        "responses": {"200": {"description": "Current daemon catalog and board generation.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PeripheralCatalog"}}}}, "409": {"$ref": "#/components/responses/Conflict"}, "502": {"$ref": "#/components/responses/BadGateway"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}, "504": {"$ref": "#/components/responses/GatewayTimeout"}}
      }
    },
    "/api/peripherals/refresh": {
      "post": {
        "tags": ["Peripherals"],
        "summary": "Refresh the peripheral catalog",
        "description": "Asks the daemon to reconcile providers, then waits until `scan_sequence` reaches its returned `target_scan_sequence`. Insight does not scan hardware itself.",
        "operationId": "refreshPeripheralCatalog",
        "requestBody": {"required": false, "content": {"application/json": {"schema": {"type": "object", "properties": {"board_generation": {"type": "integer", "minimum": 0, "description": "Generation from a previously read catalog; mismatches fail instead of refreshing another board."}}}}}},
        "responses": {"200": {"description": "Catalog produced by or after the requested reconciliation.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PeripheralCatalog"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "409": {"$ref": "#/components/responses/Conflict"}, "502": {"$ref": "#/components/responses/BadGateway"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}, "504": {"$ref": "#/components/responses/GatewayTimeout"}}
      }
    },
    "/api/peripherals/events": {
      "get": {
        "tags": ["Peripherals"],
        "summary": "Poll peripheral catalog events",
        "description": "Forwards a bounded daemon event long-poll. Use `instance_id` and `after_sequence` from the previous response; when `resync_required` is true, read the full catalog and start a new cursor.",
        "operationId": "getPeripheralEvents",
        "parameters": [
          {"name": "after_sequence", "in": "query", "schema": {"type": "integer", "minimum": 0, "default": 0}},
          {"name": "wait_ms", "in": "query", "schema": {"type": "integer", "minimum": 0, "maximum": 30000, "default": 0}},
          {"name": "instance_id", "in": "query", "schema": {"type": "string"}},
          {"name": "board_generation", "in": "query", "schema": {"type": "integer", "minimum": 0}, "description": "Optional selected-board generation guard."}
        ],
        "responses": {"200": {"description": "Daemon events or an explicit resynchronization request.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PeripheralEvents"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "409": {"$ref": "#/components/responses/Conflict"}, "502": {"$ref": "#/components/responses/BadGateway"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}, "504": {"$ref": "#/components/responses/GatewayTimeout"}}
      }
    },
...
  },
  "components": {
...
    "responses": {
...
      "BadRequest": {"description": "Missing or invalid request data.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "Conflict": {"description": "The source exists but is not currently running.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "BadGateway": {"description": "A required backend command, vf, DevKit, or remote catalog service failed or returned an invalid response.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "GatewayTimeout": {"description": "A required backend command or media helper timed out.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "ServiceUnavailable": {"description": "The selected board's required local daemon is not installed, running, or accessible.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Standard JSON error envelope returned by most failed API operations.",
        "required": ["error"],
        "properties": {"error": {"type": "string", "description": "Human-readable failure reason suitable for logs and user-facing diagnostics."}, "code": {"type": "string", "description": "Stable machine-readable error code when available."}, "hint": {"type": ["string", "null"], "description": "Actionable recovery guidance when available."}},
        "example": {"error": "Missing index"}
      },
...
      "PeripheralDevice": {
        "type": "object",
        "description": "Consumer-neutral daemon record. The object named by `type` holds provider-specific details; unknown types and optional fields must be preserved.",
        "required": ["id", "type", "provider"],
        "properties": {"id": {"type": "string"}, "type": {"type": "string"}, "provider": {"type": "string"}},
        "additionalProperties": true
      },
      "PeripheralCatalog": {
        "type": "object",
        "description": "The daemon v1 catalog plus Insight's selected-board context.",
        "required": ["schema_version", "instance_id", "state", "ready", "stale", "revision", "sequence", "scan_sequence", "devices", "board_generation", "board"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1}, "instance_id": {"type": "string"}, "state": {"type": "string", "enum": ["starting", "ready", "degraded"]}, "ready": {"type": "boolean"}, "stale": {"type": "boolean"},
          "revision": {"type": "integer", "minimum": 0}, "sequence": {"type": "integer", "minimum": 0}, "scan_sequence": {"type": "integer", "minimum": 0},
          "last_success_at": {"type": ["string", "null"], "format": "date-time"}, "last_attempt_at": {"type": ["string", "null"], "format": "date-time"}, "error": {"type": ["object", "null"], "additionalProperties": true},
          "devices": {"type": "array", "items": {"$ref": "#/components/schemas/PeripheralDevice"}}, "board_generation": {"type": "integer", "minimum": 0}, "board": {"type": "object", "required": ["label", "source"], "properties": {"label": {"type": "string"}, "source": {"type": "string"}}}
        },
        "additionalProperties": true
      },
      "PeripheralEvents": {
        "type": "object",
        "description": "The daemon v1 event envelope plus Insight's selected-board context. Event objects are passed through without local diffing.",
        "required": ["schema_version", "instance_id", "revision", "sequence", "scan_sequence", "resync_required", "shutting_down", "events", "board_generation", "board"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1}, "instance_id": {"type": "string"}, "revision": {"type": "integer", "minimum": 0}, "sequence": {"type": "integer", "minimum": 0}, "scan_sequence": {"type": "integer", "minimum": 0},
          "resync_required": {"type": "boolean"}, "shutting_down": {"type": "boolean"}, "events": {"type": "array", "items": {"type": "object", "additionalProperties": true}}, "board_generation": {"type": "integer", "minimum": 0}, "board": {"type": "object", "additionalProperties": true}
        },
        "additionalProperties": true
      },
...
    }
  }
}